Update the watcher daemon to Django v1.7
This patch removes the call to 'setup_environ()' method from the watcher
module, that has been removed from the django.core.management module in
Django v1.6, and replaces it with the 'get_wsgi_application()' one.

Also, it removes the 'RemovedInDjango18Warning' by replacing the
'close_connection' method of the 'django.db' module with the
'close_old_connections' equivalent.