Fix Admin actions in 'apply' application
Migrating to Django v1.7 and subsequently to the new migration schema,
causes the Admin model of the 'apply' app to crash since it can not
resolve the dependency on the missing 'ganeti_cluster' table anymore.

This patch fixes this issue by checking the existence of the dependent
table before creating the appropriate actions.