Improve network handling in instance application
Currently, when creating an instance we wrongly use the network's
'link' attribute as the 'network' attribute of the NIC dictionary passed
to the Ganeti RAPI client. This results in failures for instance create
operation with 'routed' NICs, unless the network name matches the link's
name. In addition, it is not possible to create instances with 'bridged'
NICs and assigning an IP address from the network's pool, just like the
routed behavior.

To overcome those issues, we use the network's description as the
'network' attribute of the 'instance_param' dictionary and we also
create a new instance_param, the network's 'link'. Finally, we add the
'ip=pool' argument to the NICs dictionary for both 'routed' and
'bridged' NICs to auto-assign an IP to an instance.