Rename 'ganeti' application to 'gnt'
In a next patch, we're going to remove the Ganeti RAPI client from
inside the GanetiMgr project (util/client.py), and use the
'python-ganeti-rapi' Debian package instead. To do so, we should first
rename the 'ganeti' application to 'gnt' to save potential importing
errors. The renaming procedure should be backwards compatible and to not
affect the current 'ganeti' model tables.

To do so, we make use of the app registry feature introduced in Django
v1.7, which stores the project's configuration and let us change various
app attributes. So, we've created a new module, the 'gnt/apps.py', where
we define the 'gnt' app attributes, and we explicitly relabel the 'gnt'
application to 'ganeti', to save renaming the model tables which would
break the project's compatibility. Finally, we explicitly make the
application to load our self-defined AppConfig within the
'gnt/__init__.py' file.